+++ changelog.md
@@ -17,6 +17,9 @@ _Nothing yet..._
 - bpk-component-tooltip: 3.0.8 => 3.1.0
   - New `popperModifiers` prop. Use this to customise the behaviour of the underlying positioning library. Please refer to its [documentation](https://github.com/FezVrasta/popper.js/blob/v1.12.9/docs/_includes/popper-documentation.md#modifiers) for details.
 
+- react-native-bpk-component-text-input:
+  - New `description` prop. Use this to add a brief description of the field that will appear underneath it.
+
 **Fixed:**
 - react-native-bpk-component-text: 2.1.28 => 2.1.29
   - Removed line height styles, allowing the built in native ones to be used instead.